+++ packages/hw-ledger/src/defaults.ts
@@ -2,17 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type Transport from '@ledgerhq/hw-transport';
-import type { SubstrateApp } from '@zondax/ledger-polkadot';
+import type { SubstrateApp } from '@zondax/ledger-substrate';
 
-import { newDockApp, newEquilibriumApp, newKusamaApp, newPolkadotApp, newPolymeshApp } from '@zondax/ledger-polkadot';
+import { newDockApp, newEdgewareApp, newEquilibriumApp, newKusamaApp, newPolkadotApp, newPolymeshApp, newStatemineApp } from '@zondax/ledger-substrate';
 
-// These match up with the network keys in the @polkadot/networks package
-// (which is turn aligns with the substrate/ss58-registry.json as the single
-// source of truth)
+// These match up with the keys of the knownLedger object in the @polkadot/networks/defaults.ts
 export const ledgerApps: Record<string, (transport: Transport) => SubstrateApp> = {
   'dock-mainnet': newDockApp,
+  edgeware: newEdgewareApp,
   equilibrium: newEquilibriumApp,
   kusama: newKusamaApp,
   polkadot: newPolkadotApp,
-  polymesh: newPolymeshApp
+  polymesh: newPolymeshApp,
+  statemine: newStatemineApp
 };