+++ packages/networks/src/defaults.ts
@@ -23,6 +23,9 @@ export const knownGenesis: KnownGenesis = {
     '0x3d75507dd46301767e601265791da1d9cb47b6ebc94e87347b635e5bf58bd047', // Snakenet Gen2
     '0x0ed32bfcab4a83517fac88f2aa7cbc2f88d3ab93be9a12b6188a036bf8a943c2' // Snakenet Gen1
   ],
+  karura: [
+    '0xbaf5aabe40646d11f0ee8abbdc64f4a4b7674925cba08e4a05ff9ebed6e2126b'
+  ],
   kulupu: [
     '0xf7a99d3cb92853d00d5275c971c132c074636256583fee53b3bbe60d7b8769ba'
   ],
@@ -40,6 +43,9 @@ export const knownGenesis: KnownGenesis = {
   polymesh: [
     '0x9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f'
   ],
+  statemine: [
+    '0x48239ef607d7928874027a43a67689209727dfb3d3dc5e5b03a39bdc2eda771a'
+  ],
   subsocial: [
     '0x0bd72c1c305172e1275278aaeb3f161e02eccb7a819e63f62d47bd53a28189f8'
   ]
@@ -48,16 +54,23 @@ export const knownGenesis: KnownGenesis = {
 // these are icon overrides
 export const knownIcon: KnownIcon = {
   kusama: 'polkadot',
-  polkadot: 'polkadot'
+  polkadot: 'polkadot',
+  statemine: 'polkadot',
+  statemint: 'polkadot',
+  westmint: 'polkadot'
 };
 
-// support for ledger
+// These match up with the keys of the ledgerApps object in the @polkadot/hw-ledger/defaults.ts
+// and maps to the known slip44 (minus the `0x8` hard derivation flag)
+// NOTE: Any network here needs to have a genesisHash attached in the config above
 export const knownLedger: KnownLedger = {
   'dock-mainnet': 0x00000252,
+  edgeware: 0x0000020b,
   equilibrium: 0x05f5e0fd,
   kusama: 0x000001b2,
   polkadot: 0x00000162,
-  polymesh: 0x00000253
+  polymesh: 0x00000253,
+  statemine: 0x000001b2 // common-good on Kusama, shares derivation
 };
 
 // testnets should not allow selection