+++ packages/util-crypto/src/base64/validate.ts
@@ -10,7 +10,7 @@ import { assert } from '@polkadot/util';
  * Validates that the supplied value is valid base64
  */
 export function base64Validate (value?: string | null): value is string {
-  assert(value, 'Expected non-null, non-empty base64 input');
+  assert(value && typeof value === 'string', 'Expected non-null, non-empty base64 string input');
   assert(/^(?:[A-Za-z0-9+/]{2}[A-Za-z0-9+/]{2})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$/.test(value), 'Invalid base64 encoding');
 
   return true;