+++ packages/util/src/array/index.ts
@@ -8,3 +8,5 @@
 export { arrayChunk } from './chunk';
 export { arrayFilter } from './filter';
 export { arrayFlatten } from './flatten';
+export { arrayRange } from './range';
+export { arrayShuffle } from './shuffle';