+++ packages/util/src/bn/consts.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import BN from 'bn.js';
+import { BN } from './bn';
 
 /**
  * @name BN_ZERO