+++ packages/util/src/bn/sqrt.ts
@@ -3,10 +3,9 @@
 
 import type { ToBn } from '../types';
 
-import BN from 'bn.js';
-
 import { assert } from '../assert';
-import { BN_MAX_INTEGER, BN_ONE, BN_ZERO } from '../bn';
+import { BN } from './bn';
+import { BN_MAX_INTEGER, BN_ONE, BN_ZERO } from './consts';
 import { bnToBn } from './toBn';
 
 const SQRT_MAX_SAFE_INTEGER = new BN(94906265);