+++ packages/util/src/bn/toU8a.ts
@@ -1,8 +1,8 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type BN from 'bn.js';
 import type { ToBn, ToBnOptions } from '../types';
+import type { BN } from './bn';
 
 import { isNumber } from '../is/number';
 import { bnToBn } from './toBn';