+++ packages/util/src/compact/toU8a.spec.ts
@@ -1,8 +1,7 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import BN from 'bn.js';
-
+import { BN } from '../bn';
 import { compactToU8a } from '.';
 
 describe('encode', (): void => {