+++ packages/util/src/format/formatElapsed.ts
@@ -1,7 +1,7 @@
-// Copyright 2017-2021 @polkadot/ui-util authors & contributors
+// Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type BN from 'bn.js';
+import type { BN } from '../bn/bn';
 import type { ToBn } from '../types';
 
 import { bnToBn } from '../bn/toBn';